remove unused backup files, reformat listing page to not include dynamic resize panels
